kernel clean up

+++ resources-ext/opencl_kernels/dataprocessing/floatDiv.cl
@@ -1,6 +1,6 @@
 __constant sampler_t sampler = CLK_NORMALIZED_COORDS_FALSE | CLK_ADDRESS_NONE | CLK_FILTER_NEAREST;
 
-__kernel void floatDiv(__global unsigned int* divider, read_only image2d_t matrix, write_only image2d_t matrix_write) {
+__kernel void floatDiv(__global float* divider, read_only image2d_t matrix, write_only image2d_t matrix_write) {
     int gidx = get_global_id(0);
     int gidy = get_global_id(1);
